+++ src/main/resources/static/js/editor.js
@@ -1,45 +1,35 @@
-
-  window.editorInstance = null;
-
-  window.initEditor = function(targetSelector, content = '') {
-    if (window.editorInstance) {
-      window.editorInstance.destroy();
-      window.editorInstance = null;
-    }
-
-    window.editorInstance = new toastui.Editor({
-      el: document.querySelector(targetSelector),
-      height: '100%',
-      initialEditType: 'wysiwyg',
-      previewStyle: 'vertical',
-      initialValue: content,
-      hideModeSwitch: true,
-      toolbarItems: [
-        ['heading', 'bold', 'italic'],
-        ['ul', 'ol']
-      ],
-      hooks: {
-        addImageBlobHook: () => {
-          alert('이미지는 업로드할 수 없습니다.');
-          return false;
-        }
-      }
-    });
-
-    return window.editorInstance;
-  };
-  
-  
-  let editorLoaded = false;
-  export function loadEditorScript(callback) {
-      if (editorLoaded) return callback();
-
-      const script = document.createElement('script');
-      script.src = '/SOLEX/js/editor.js';
-      script.onload = () => {
-          editorLoaded = true;
-          callback();
-      };
-
-      document.body.appendChild(script);
+function initEditor(selector, content = '') {
+  // 기존 인스턴스 제거 (모달 재열기 시 중복 방지)
+  if (window.editorInstance) {
+    window.editorInstance.destroy();  // 기존 인스턴스 정리
+    window.editorInstance = null;
   }
+
+  const el = document.querySelector(selector);
+  if (!el) {
+    console.error('Editor element not found:', selector);
+    return;
+  }
+
+  window.editorInstance = new toastui.Editor({
+    el,
+    height: '500px',
+    initialEditType: 'wysiwyg',
+    previewStyle: 'vertical',
+    initialValue: content,
+    toolbarItems: [
+      ['heading', 'bold', 'italic'],
+      ['ul', 'ol']
+    ],
+    hideModeSwitch: true,
+    hooks: {
+      addImageBlobHook: () => {
+        alert('이미지는 업로드할 수 없습니다.');
+        return false;
+      }
+    }
+  });
+}
+
+// 전역 등록
+window.initEditor = initEditor;
