+++ src/main/resources/templates/notice/noticeList.html
@@ -8,7 +8,6 @@
 <!-- <meta charset="UTF-8"> -->
 <title>SOLEX</title>
 <link rel="stylesheet" th:href="@{/css/notice/notice.css}" />
-
 <style>
 </style>
 </head>
@@ -49,10 +48,7 @@
 		</div>
 		</div>
 		
-		<!--<script type="module" th:src="@{/js/notice/notice.js}" defer></script>-->
-		<!-- Script: 순서 중요 -->
-		<script type="module" th:src="@{/js/notice/modal.js}"></script>
-		<script type="module" th:src="@{/js/notice/notice.js}"></script>
+		<script th:src="@{/js/notice/notice.js}"></script>
 		
 	</th:block>
